update: script

+++ setuprepo.sh
@@ -1,19 +1,20 @@
 
-
 cd "$(dirname "${BASH_SOURCE:-$0}")"
 
 repo init \
-    -u https://github.com/LineageOS/android.git \
-    -b lineage-18.1
+  -u https://github.com/LineageOS/android.git \
+  -b lineage-18.1
 
 git clone \
   -b lineage-18.1 \
   https://github.com/lin18-microG/local_manifests.git \
   .repo/local_manifests
 
-mv manifest.xml .repo/local_manifests/setup_sony.xml
+mv \
+  manifest.xml \
+  .repo/local_manifests/setup_sony.xml
 
-rm -rf .repo/local_manifests/.git
-rm -rf .git
-rm README.md
-rm "${BASH_SOURCE:-$0}"
+[ -d .repo/local_manifests/.git ] && rm -rf .repo/local_manifests/.git
+[ -d .git                       ] && rm -rf .git
+[ -f README.md                  ] && rm README.md
+[ -f "${BASH_SOURCE:-$0}"       ] && rm "${BASH_SOURCE:-$0}"
